For local charms, use absolute path.
This is dependent on juju-deployer first supporting charms at absolute
paths.

Instead of making a copy of the local charm and initing it as a new bzr
repo, leave the charm in place and tell deployer where it is. A temp
copy will still be made if we need a proper series parent directory, but
it is just a dir copy and not a bzr repo.

The implications of this change are that the charm under test can be
modified in place, and amulet/deployer will always see (and test) the
latest version of that charm as it exists on disk. Changes need not be
committed first, and in fact, the charm under test does not even need to
be under revision control.