Fix sentry tools no longer defaulting to charm dir
With the change to use 'juju ssh' in lieu of 'juju run' for the sentry
helpers, the default working directory is no longer the charm dir.  Work
around this by explicitly changing the directory (though not for
juju_agent, since the explicit charm dir interferes with it watching for
hooks).